Added 2 Examples to the ReadMe
1) Iterating though search results

The documentation does not seem to cover how to iterate through objects. Many developers perform a var_dump(), don't see the object data, and assume they need to then independently lookup each with a find request. This is not true.

2) Finding all customers funding options.

Many developer will want to display to a customer a custom list of funding options. It is not correctly shown in the documentation you must use $customer->paymentMethods() as a function to get all. It is not shown that you can use isDefault() to check if this source is the default. It is also not obvious how to present a consistent label for the field as you loop though.