Fix fatal error 'class Braintree_Modification not found'
[GitHub issue #64](https://github.com/braintree/braintree_php/issues/64)
reports that this fatal error occurs when using the Composer autoloader
instead of requiring lib/Braintree.php. The error occurs because of the
following sequence of events:

1. The composer autoloader includes lib/Braintree/Modification.php and
begins parsing the file.
2. The class Braintree is not found, so it is autoloaded.
lib/Braintree.php is loaded and parsed.
3. lib/Braintree.php contains a require_once(lib/Braintree/Modification.php)
which has no effect, since that file was already included
4. lib/Braintree.php contains a require_once(lib/Braintree/AddOn.php)
which loads and parses that file.
5. Braintree_AddOn extends Braintree_Modification. While the file that
defines Braintree_Modification was loaded, it has not completely been
evaluated at this point, and so Braintree_Modification has not been
defined. lib/Braintree/Modification.php has only been evaluated up until
`extends Braintree` (step 2).

It is considered [best practice in PHP][PSR-1] that files should either
declare symbols (e.g. classes) or cause side effects (e.g. requiring all
of the files for the library) but not both. This PR moves the Braintree
class out of lib/Braintree.php and instead defines it as Braintree_Base.
This approach was chosen because our current documentation directs users
to bootstrap the library by including lib/Braintree.php. This PR does
not break that interface.

In addition, a sanity test is added to assert that the library can be
used via the Composer autoloader. This test only works on PHP 5.4 or
higher because of its reliance on the `PHP_BINARY` constant. We recently
deprecated support for PHP 5.2 and 5.3, and we use PHP 5.4 on CI, so
this should not be a problem.

[PSR-1]:
https://github.com/php-fig/fig-standards/blob/558a30cb5c6e411d3b984897ba9ef2dbff16e519/accepted/PSR-1-basic-coding-standard.md#1-overview