Format DateTime objects before converting them to strings
Fixes [GitHub #71](https://github.com/braintree/braintree_php/issues/71)

When calling __toString() on objects that inherit from
Braintree_Instance, we concatenate all of the attributes and their
values to create string. This relies on values being castable to a
string. However, DateTime objects do not implement __toString() and as
such cannot be cast to a string.

This change explicitly formats DateTime instances to an RFC 2822
date/time string when calling Braintree_Instance::__toString().