BTCOMINF-220 - Remove `signature` from OAuth connect URL
This removes unneeded `signature` and `algorithm` parameters from the
OAuth connect URL generation. It is no longer needed.

The method that generates this URL calls out to another method, which
is deprecated as of this change. This is the only change noted in the
changelog.

`rake test:php:unit && rake test:single[OAuthTest]` passes after this
change.