BTCARDSXP-372 Line item discount, unit tax, and tax amount may be zero   (#167)
What
===
Update the validations for transaction line item discount amount, tax amount, and unit tax amount to indicate that a zero value is allowed.

Why
===
The discount amount and tax amount can be zero and we accidentally wrote a validation that disallowed it. The same change has been made to unit tax amount for consistency, @abuttars is the significant owner of that field and he 👍 d allowing it there too.