BTCARDSXP-381 Move Transaction\LineItem => TransactionLineItem (#168)
What
===
Move the `Braintree\Transaction\LineItem` class out of the `Transaction` namespace to `Braintree\TransactionLineItem`, and alias the old class to the new one.

Why
=== 
Transaction line item is defined as a top level class in every other SDK and in the docs. It's a class that has its own gateway, transaction line item gateway, and is not a class that is part of the transaction response like the other classes that are in the transaction namespace.

Additionally, the docs apps renders the class as not being in the namespace in its docs.

To bring this SDK into consistency with the others and with the docs it is necessary to move this class.

The use of `class_alias` is so that the original class is still a valid class and both classes point to the same definition, for backwards compatibility. The `class_alias` function is available in PHP 5.3+ and the PHP SDK requires a minimum of PHP 5.4.

Notes
===

After moving and alias the class I ran all the tests without changing any of the references to the old class in both the lib and tests. All relevant tests passed. I then made the change to the lib and tests to use the new class and ran again. All relevant tests passed. I've added two small tests to check that the enumerations in the old class continue to be referenced for backwards compatibility.