BTCARDSXP-530 Fix transaction source test expectation (#195)
### What
Change the expectation in the transaction source is invalid validation
error test to get the code from the first field error on the
`transactionSource` attribute.

### Why
It was attempting to get the code from the array of field errors, which
was a mistake. There are multiple field errors for each attribute and we
want to get the first error and check it matches what we expect.