BTPWPP-1550 - Support creating PayPal transactions with local payment method (#216)
webhook content

### What
Add support for creating transactions with payment ID and payer ID,
which come from a local payment method webhook.

### Why
During the PayPal Checkout flow, a customer may complete a payment but
not return to the merchant's page, meaning that the merchant would never
generate a nonce. However, the merchant can subscribe to local payment
method webhooks, in which case they would receive a webhook notifying
them of the payment. In order to import that transaction to their
gateway, they will need to create a transaction using that webhook
content.